Rename folders to remove "questionnaire2"
