+++ .config/configstore/update-notifier-nodemon.json
@@ -1,0 +1,4 @@
+{
+	"optOut": false,
+	"lastUpdateCheck": 1557318059222
+}
