--- questionnaire/questionnaire.njk
@@ -1,178 +1,0 @@
-{% extends "page.njk" %}
-
-{% block content %}
-    {% from "radios/macro.njk" import govukRadios %}
-{% from "input/macro.njk" import govukInput %}
-{% from "input/macro.njk" import govukInput %}
-{% from "input/macro.njk" import govukInput %}
-{% from "input/macro.njk" import govukInput %}
-{% from "input/macro.njk" import govukInput %}
-{% from "input/macro.njk" import govukInput %}
-{% from "input/macro.njk" import govukInput %}
-
-{% set emailHtml %}
-{{ govukInput({
-  id: "contact-by-email",
-  name: "contact-by-email",
-  type: "email",
-  classes: "govuk-!-width-one-third",
-  label: {
-    text: "Email address"
-  }
-}) }}
-{% endset -%}
-
-{% set phoneHtml %}
-{{ govukInput({
-  id: "contact-by-phone",
-  name: "contact-by-phone",
-  type: "tel",
-  classes: "govuk-!-width-one-third",
-  label: {
-    text: "Phone number"
-  }
-}) }}
-{% endset -%}
-
-{% set textHtml %}
-
-{% endset -%}
-
-{{ govukRadios({
-  idPrefix: "how-contacted-conditional",
-  name: "how-contacted",
-  fieldset: {
-    legend: {
-      "text": "How would you prefer to be contacted? hhhhhhhhhhhhhhhhhhhhhhhh",
-      isPageHeading: true,
-      classes: "govuk-fieldset__legend--xl"
-    }
-  },
-  hint: {
-    text: "Select one option."
-  },
-  items: [
-    {
-      value: "email",
-      text: "Email",
-      conditional: {
-        html: emailHtml
-      }
-    },
-    {
-      value: "phone",
-      text: "Phone",
-      conditional: {
-        html: phoneHtml
-      }
-    },
-    {
-      value: "text",
-      text: "Text message",
-      conditional: {
-        html: govukInput({
-            id: "contact-by-text",
-            name: "contact-by-text",
-            type: "tel",
-            classes: "govuk-!-width-one-third",
-            label: {
-                text: "Mobile phone number zzzz"
-            }
-        })
-      }
-    }
-  ]
-}) }}
-
-
-
-{{ govukRadios({
-    classes: "govuk-radios--inline",
-    idPrefix: "changed-name",
-    name: "changed-name",
-    fieldset: {
-        legend: {
-            text: "Have you changed your name?",
-            isPageHeading: true,
-            classes: "govuk-fieldset__legend--xl"
-        }
-    },
-    hint: null,
-    items: [
-        {
-            value: "yes",
-            text: "Yes"
-        },
-        {
-            value: "no",
-            text: "No"
-        }
-    ]
-}) }}
-
-
-
-{#
-{% from "date-input/macro.njk" import govukDateInput %}
-
-{{ govukDateInput({
-  id: "dob",
-  namePrefix: "dob",
-  fieldset: {
-    legend: {
-      text: "What is your date of birth?",
-      isPageHeading: true,
-      classes: "govuk-fieldset__legend--xl"
-    }
-  },
-  hint: {
-    text: "For example, 31 3 1980"
-  }
-}) }}
-#}
-
-
-{% set dayzzz %}
-    {{ govukInput({
-    label: {
-        text: "Dayzzz"
-    },
-    id: "dayzzz",
-    name: "day"
-    }) }}
-{% endset -%}
-
-
-{% from "date-input/macro.njk" import govukDateInput %}
-
-{{ govukDateInput({
-  id: "dob",
-  namePrefix: "dob",
-  fieldset: {
-    legend: {
-      text: "What is your date of birth?",
-      isPageHeading: true,
-      classes: "govuk-fieldset__legend--xl"
-    }
-  },
-  hint: {
-    text: "For example, 31 3 1980"
-  },
-  items: [
-    dayzzz,
-    {
-      classes: "govuk-input--width-2",
-      name: "month",
-      value: "3"
-    },
-    {
-      classes: "govuk-input--width-4",
-      name: "year",
-      value: "2076"
-    }
-  ]
-}) }}
-
-
-
-{% endblock %}
